Make prompt text of consult-denote-file-prompt be consistent with denote-file-prompt
This is in light of commit dda6c5a in denote.git that I just made there.

The idea is that if the user has configured denote-directory to a list
of directories, then we want to inform them where the file prompt is
at. For a single directory, this is superfluous.